+++ src/components/icons/CalendarCheckIcon.tsx
@@ -1,0 +1,5 @@
+export const CalendarCheckIcon = () => (
+  <svg width="25" height="23" viewBox="-0.5 -0.3 3 3" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M0.854634 1.61344L1.64742 0.751716L1.48999 0.580594L0.854634 1.27119L0.534146 0.922838L0.376714 1.09396L0.854634 1.61344ZM0 2.20014V0H2.02413V2.20014H0ZM0.224904 1.95568H1.79923V0.24446H0.224904V1.95568Z" fill="#973BEB" />
+  </svg>
+);
