+++ src/components/icons/CalendarIcon.tsx
@@ -1,0 +1,6 @@
+export const CalendarIcon = () => (
+  <svg width="25" height="25" viewBox="-0.2 0 4 4" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M3.21006 1.45835V0.729187H0.458557V1.45835M3.21006 1.45835H0.458557M3.21006 1.45835V3.06252H0.458557V1.45835M1.07 0.729187V0.291687M2.59862 0.729187V0.291687" stroke="#6059E8" stroke-width="0.25" stroke-linecap="square"/>
+    <path d="M2.36929 1.93231L1.7207 2.55108L1.39648 2.24162" stroke="#6059E8" stroke-width="0.25" stroke-linecap="square"/>
+  </svg>
+);
