+++ src/components/icons/ChevronDownIcon.tsx
@@ -1,0 +1,5 @@
+export const ChevronDownIcon = () => (
+  <svg width="25" height="20" viewBox="0 -0.2 4 2" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M1.4303 1.08622L0.713559 0.481237L0.892712 0.330017L1.51988 0.859392L2.14704 0.330017L2.32619 0.481237L1.60945 1.08622C1.58569 1.10627 1.55347 1.11753 1.51988 1.11753C1.48628 1.11753 1.45406 1.10627 1.4303 1.08622Z" fill="#373737" />
+  </svg>
+);
