+++ src/components/icons/MapPinIcon.tsx
@@ -1,0 +1,6 @@
+export const MapPinIcon = () => (
+  <svg width="20" height="20" viewBox="-0.7 -0.4 2 3" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M1.73049 0.896275C1.73049 1.54327 0.915234 2.09072 0.915234 2.09072C0.915234 2.09072 0.0999756 1.54327 0.0999756 0.896275C0.0999756 0.685083 0.185869 0.482541 0.338759 0.333206C0.49165 0.183871 0.699014 0.0999756 0.915234 0.0999756C1.13145 0.0999756 1.33882 0.183871 1.49171 0.333206C1.6446 0.482541 1.73049 0.685083 1.73049 0.896275Z" stroke="#373737" stroke-width="0.2" />
+    <path d="M1.22094 0.896268C1.22094 0.975465 1.18873 1.05142 1.1314 1.10742C1.07406 1.16342 0.996302 1.19488 0.915219 1.19488C0.834137 1.19488 0.756375 1.16342 0.699041 1.10742C0.641707 1.05142 0.609497 0.975465 0.609497 0.896268C0.609497 0.817071 0.641707 0.741118 0.699041 0.685118C0.756375 0.629117 0.834137 0.597656 0.915219 0.597656C0.996302 0.597656 1.07406 0.629117 1.1314 0.685118C1.18873 0.741118 1.22094 0.817071 1.22094 0.896268Z" stroke="#373737" stroke-width="0.2" />
+  </svg>
+);
