+++ src/components/icons/PercentIcon.tsx
@@ -1,0 +1,5 @@
+export const PercentIcon = () => (
+  <svg width="25" height="25" viewBox="-0.2 -0.2 4 4" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M1.1001 1.61347C0.958311 1.61347 0.837303 1.56335 0.737074 1.46312C0.636845 1.36289 0.586731 1.24189 0.586731 1.1001C0.586731 0.958311 0.636845 0.837303 0.737074 0.737074C0.837303 0.636845 0.958311 0.586731 1.1001 0.586731C1.24189 0.586731 1.36289 0.636845 1.46312 0.737074C1.56335 0.837303 1.61347 0.958311 1.61347 1.1001C1.61347 1.24189 1.56335 1.36289 1.46312 1.46312C1.36289 1.56335 1.24189 1.61347 1.1001 1.61347ZM1.1001 1.32011C1.16121 1.32011 1.21319 1.29875 1.25601 1.25601C1.29884 1.21328 1.32021 1.16131 1.32011 1.1001C1.32001 1.03889 1.29865 0.986962 1.25601 0.944328C1.21338 0.901694 1.16141 0.880279 1.1001 0.880084C1.03879 0.879888 0.986864 0.901303 0.944328 0.944328C0.901792 0.987353 0.880377 1.03928 0.880084 1.1001C0.87979 1.16092 0.901205 1.21289 0.944328 1.25601C0.987451 1.29914 1.03937 1.3205 1.1001 1.32011ZM2.42018 2.93355C2.2784 2.93355 2.15739 2.88344 2.05716 2.78321C1.95693 2.68298 1.90682 2.56197 1.90682 2.42018C1.90682 2.2784 1.95693 2.15739 2.05716 2.05716C2.15739 1.95693 2.2784 1.90682 2.42018 1.90682C2.56197 1.90682 2.68298 1.95693 2.78321 2.05716C2.88344 2.15739 2.93355 2.2784 2.93355 2.42018C2.93355 2.56197 2.88344 2.68298 2.78321 2.78321C2.68298 2.88344 2.56197 2.93355 2.42018 2.93355ZM2.42018 2.6402C2.4813 2.6402 2.53327 2.61883 2.5761 2.5761C2.61893 2.53337 2.6403 2.4814 2.6402 2.42018C2.6401 2.35897 2.61874 2.30705 2.5761 2.26441C2.53347 2.22178 2.4815 2.20037 2.42018 2.20017C2.35887 2.19997 2.30695 2.22139 2.26441 2.26441C2.22188 2.30744 2.20046 2.35936 2.20017 2.42018C2.19988 2.48101 2.22129 2.53298 2.26441 2.5761C2.30754 2.61922 2.35946 2.64059 2.42018 2.6402ZM0.689404 2.83088C0.662514 2.80399 0.649068 2.76976 0.649068 2.72821C0.649068 2.68665 0.662514 2.65242 0.689404 2.62553L2.62553 0.689404C2.65242 0.662514 2.68665 0.649068 2.7282 0.649068C2.76976 0.649068 2.80399 0.662514 2.83088 0.689404C2.85777 0.716295 2.87121 0.75052 2.87121 0.792078C2.87121 0.833636 2.85777 0.867861 2.83088 0.894751L0.894751 2.83088C0.867861 2.85777 0.833636 2.87121 0.792078 2.87121C0.75052 2.87121 0.716295 2.85777 0.689404 2.83088Z" fill="#EB5910" />
+  </svg>
+);
