+++ src/components/icons/SelectCheckIcon.tsx
@@ -1,0 +1,5 @@
+export const SelectCheckIcon = () => (
+  <svg width="25" height="20" viewBox="0 0 4 4" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M2.5833 0.942071C2.61031 0.967477 2.62744 1.00418 2.63092 1.0441C2.6344 1.08403 2.62394 1.12391 2.60186 1.15497L1.41765 2.81945C1.40601 2.83581 1.39152 2.84918 1.37511 2.85871C1.3587 2.86824 1.34074 2.8737 1.32238 2.87476C1.30402 2.87581 1.28567 2.87243 1.26851 2.86484C1.25136 2.85724 1.23578 2.84559 1.22278 2.83064L0.564889 2.07406C0.552322 2.06011 0.542298 2.04341 0.535402 2.02495C0.528506 2.00649 0.524876 1.98663 0.524724 1.96654C0.524572 1.94645 0.527902 1.92652 0.534517 1.90793C0.541133 1.88933 0.550903 1.87244 0.563258 1.85823C0.575612 1.84402 0.590303 1.83279 0.606474 1.82518C0.622644 1.81757 0.63997 1.81374 0.657441 1.81391C0.674912 1.81409 0.692178 1.81826 0.708231 1.82619C0.724284 1.83412 0.738803 1.84565 0.750941 1.8601L1.30607 2.4985L2.39817 0.963406C2.42027 0.932349 2.45218 0.912654 2.4869 0.908653C2.52161 0.904652 2.55629 0.916672 2.5833 0.942071Z" fill="white" />
+  </svg>
+);
