+++ src/components/icons/SmallUsersIcon.tsx
@@ -1,0 +1,5 @@
+export const SmallUsersIcon = () => (
+  <svg width="22" height="22" viewBox="-0.5 -0.4 4 3" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M1.02078 1.46209C1.31476 1.46209 1.56791 1.51436 1.75063 1.63941C1.93989 1.76903 2.04155 1.96874 2.04155 2.22943C2.03991 2.2558 2.02844 2.28054 2.00948 2.29862C1.99052 2.31671 1.96549 2.32677 1.93948 2.32677C1.91347 2.32677 1.88844 2.31671 1.86947 2.29862C1.85051 2.28054 1.83904 2.2558 1.8374 2.22943C1.8374 2.03075 1.76431 1.89844 1.63671 1.81113C1.50238 1.71946 1.29639 1.66948 1.02078 1.66948C0.745168 1.66948 0.538971 1.71946 0.40484 1.81134C0.277243 1.89865 0.204156 2.03096 0.204156 2.22964C0.202513 2.25601 0.191044 2.28075 0.172081 2.29883C0.153118 2.31691 0.128086 2.32698 0.102078 2.32698C0.076069 2.32698 0.0510372 2.31691 0.0320744 2.29883C0.0131117 2.28075 0.00164247 2.25601 0 2.22964C0 1.96874 0.101465 1.76903 0.290717 1.63941C0.473437 1.51436 0.726794 1.46209 1.02078 1.46209ZM2.03645 1.46209C2.33023 1.46209 2.58359 1.51436 2.76631 1.63941C2.95556 1.76903 3.05723 1.96874 3.05723 2.22964C3.05723 2.25714 3.04647 2.28352 3.02733 2.30296C3.00819 2.32241 2.98222 2.33333 2.95515 2.33333C2.92808 2.33333 2.90211 2.32241 2.88297 2.30296C2.86383 2.28352 2.85307 2.25714 2.85307 2.22964C2.85307 2.03096 2.77999 1.89865 2.65218 1.81134C2.54643 1.73896 2.39638 1.69312 2.20263 1.67653C2.15437 1.59645 2.0945 1.52423 2.02502 1.4623L2.03645 1.46209ZM1.02078 3.01627e-06C1.18853 0.00429727 1.34799 0.0750088 1.46514 0.197053C1.5823 0.319098 1.64787 0.482814 1.64787 0.653277C1.64787 0.823741 1.5823 0.987457 1.46514 1.1095C1.34799 1.23155 1.18853 1.30226 1.02078 1.30655C0.850219 1.30655 0.686647 1.23772 0.566044 1.11521C0.445442 0.992699 0.377688 0.826537 0.377688 0.653277C0.377688 0.480018 0.445442 0.313855 0.566044 0.191343C0.686647 0.0688299 0.850219 3.01627e-06 1.02078 3.01627e-06ZM2.03645 3.01627e-06C2.20701 3.01627e-06 2.37058 0.0688299 2.49118 0.191343C2.61179 0.313855 2.67954 0.480018 2.67954 0.653277C2.67954 0.826537 2.61179 0.992699 2.49118 1.11521C2.37058 1.23772 2.20701 1.30655 2.03645 1.30655C1.92947 1.30655 1.82903 1.27959 1.74043 1.23272C1.78315 1.17815 1.81972 1.11888 1.84944 1.05603C1.91635 1.08818 1.99016 1.10262 2.06403 1.09801C2.13791 1.09341 2.20943 1.0699 2.27196 1.02968C2.33449 0.989465 2.38599 0.933837 2.42168 0.867973C2.45737 0.802109 2.47608 0.728148 2.47608 0.652966C2.47608 0.577784 2.45737 0.503823 2.42168 0.437959C2.38599 0.372095 2.33449 0.316468 2.27196 0.276248C2.20943 0.236029 2.13791 0.212524 2.06403 0.207918C1.99016 0.203313 1.91635 0.217756 1.84944 0.249906C1.81963 0.18719 1.78299 0.128062 1.74022 0.073626C1.83152 0.024956 1.93314 -0.000317787 2.03625 3.01627e-06M1.02078 0.207392C0.906682 0.210927 0.798426 0.259454 0.71895 0.342688C0.639475 0.425922 0.595028 0.537321 0.595028 0.653277C0.595028 0.769234 0.639475 0.880633 0.71895 0.963867C0.798426 1.0471 0.906682 1.09563 1.02078 1.09916C1.13719 1.09916 1.24883 1.05219 1.33115 0.968566C1.41347 0.884946 1.45971 0.771534 1.45971 0.653277C1.45971 0.535021 1.41347 0.421608 1.33115 0.337989C1.24883 0.254369 1.13719 0.207392 1.02078 0.207392Z" fill="#373737" />
+  </svg>
+);
