+++ src/components/icons/TimeClockIcon.tsx
@@ -1,0 +1,5 @@
+export const TimeClockIcon = () => (
+  <svg width="15" height="15" viewBox="0 0 2 2" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M1.90935 0.811848V0.351326H0.0750122V0.811848M1.90935 0.811848H0.0750122M1.90935 0.811848V1.825H0.0750122V0.811848M0.482643 0.351326V0.0750122M1.50172 0.351326V0.0750122" stroke="#373737" stroke-width="0.15" stroke-linecap="square" />
+  </svg>
+);
