+++ src/constants/mockData.ts
@@ -1,0 +1,34 @@
+import { Meeting } from '../types/history';
+
+export const DUMMY_MEETINGS: Meeting[] = [
+  {
+    id: 1,
+    title: '프로젝트 팀',
+    date: '2024년 1월 15일',
+    time: '목요일 18:00-20:00',
+    location: '강남역 2번 출구 스타벅스',
+    participants: '6명 참석',
+    status: '100% 완료',
+    statusClasses: 'bg-green-100 text-green-700',
+  },
+  {
+    id: 2,
+    title: '독서 모임',
+    date: '2024년 1월 8일',
+    time: '토요일 14:00-16:00',
+    location: "홍대 북카페 '책과 커피'",
+    participants: '4명 참석',
+    status: '87% 완료',
+    statusClasses: 'bg-pink-100 text-red-700',
+  },
+  {
+    id: 3,
+    title: '운동 모임',
+    date: '2024년 1월 1일',
+    time: '월요일 10:00-12:00',
+    location: '한강공원 뚝섬유원지',
+    participants: '3명 참석',
+    status: '100% 완료',
+    statusClasses: 'bg-green-100 text-green-700',
+  },
+];
